Create and delete buttons #1
